refact: authenticated user 일괄 적용 및 User domain과의 결합도 감소

+++ src/main/java/com/cherry/cherrybookerbe/mylib/query/dto/request/MyLibrarySearchRequest.java
@@ -8,13 +8,13 @@ public class MyLibrarySearchRequest {
 
     private static final int DEFAULT_SIZE = 8;
 
-    private final Long userId;
+    private final Integer userId;
     private final String keyword;
     private final BookStatus status;
     private final int page;
     private final int size;
 
-    public MyLibrarySearchRequest(Long userId,
+    public MyLibrarySearchRequest(Integer userId,
                                   String keyword,
                                   BookStatus status,
                                   int page,
